# Force inline small functions used by LZ4_compress_generic.