@Cyan4973 2020-01-24
Is the speed impact mostly visible with Visual Studio ?