@wolfpld 2020-01-25
GCC 7.4.0 goes crazy and completely reorganizes code, even though the inlined functions are not seen called in neither version of the code.

```
 dickens :
Compression functions :
 1-LZ4_compress_default            :  10192446 ->  6431106 (63.10%),  381.0 MB/s -> 383.2 MB/s
 2-LZ4_compress_default(small dst) :  10192446 ->  6431106 (63.10%),  358.8 MB/s -> 383.2 MB/s
 3-LZ4_compress_destSize           :  10192446 ->  6431106 (63.10%),  401.2 MB/s -> 387.3 MB/s
 4-LZ4_compress_fast(0)            :  10192446 ->  6431106 (63.10%),  391.4 MB/s -> 381.0 MB/s
 5-LZ4_compress_fast(1)            :  10192446 ->  6431106 (63.10%),  383.2 MB/s -> 379.2 MB/s
 6-LZ4_compress_fast(2)            :  10192446 ->  6763934 (66.36%),  425.5 MB/s -> 413.2 MB/s
 7-LZ4_compress_fast(17)           :  10192446 ->  9222511 (90.48%), 1047.8 MB/s -> 1076.3 MB/s
 8-LZ4_compress_fast_extState(0)   :  10192446 ->  6431106 (63.10%),  387.3 MB/s -> 384.8 MB/s
 9-LZ4_compress_fast_continue(0)   :  10192446 ->  6428753 (63.07%),  395.5 MB/s -> 397.2 MB/s
```

Clang 8.0.1 generates identical binaries for both versions of the code.

```
 dickens :
Compression functions :
 1-LZ4_compress_default            :  10192446 ->  6431106 (63.10%),  368.6 MB/s -> 354.7 MB/s
 2-LZ4_compress_default(small dst) :  10192446 ->  6431106 (63.10%),  375.1 MB/s -> 366.9 MB/s
 3-LZ4_compress_destSize           :  10192446 ->  6431106 (63.10%),  364.7 MB/s -> 362.9 MB/s
 4-LZ4_compress_fast(0)            :  10192446 ->  6431106 (63.10%),  358.8 MB/s -> 371.0 MB/s
 5-LZ4_compress_fast(1)            :  10192446 ->  6431106 (63.10%),  368.6 MB/s -> 371.0 MB/s
 6-LZ4_compress_fast(2)            :  10192446 ->  6763934 (66.36%),  384.8 MB/s -> 403.6 MB/s
 7-LZ4_compress_fast(17)           :  10192446 ->  9222511 (90.48%), 1002.9 MB/s -> 1002.9 MB/s
 8-LZ4_compress_fast_extState(0)   :  10192446 ->  6431106 (63.10%),  344.3 MB/s -> 376.7 MB/s
 9-LZ4_compress_fast_continue(0)   :  10192446 ->  6428753 (63.07%),  366.9 MB/s -> 364.6 MB/s
```

It looks like there is a lot of noise in the results and it would be interesting to see if timings performed on an Intel CPU would be more stable.

The Odroid C2 test in the initial post was made on a "simple" in-order CPU, running at a set frequency. Even though GCC 9.2.0 was used there, you can see there was a small performance uplift in all but one tests.