@Cyan4973 2020-01-27
This always felt like a straightforward patch, but I wanted to ensure that there was no unexpected regression.

I confirm that `clang` produces the exact same binary, with or without this patch.
`gcc` changes are more random, but the impact feels within the noise margin.

On Visual though, it's a very clear win. Curiously, I measured even higher gains than claimed in this PR. In either case, it's good enough for a merge.

Thanks @wolfpld !